+++ WebContent/js/components/JobFile.js
@@ -17,7 +17,7 @@ import Description from '@material-ui/icons/Description';
 import { hideMenu } from 'react-contextmenu/modules/actions';
 import { encodeURLComponent } from '../utilities/urlUtils';
 import { fetchJobFile, getFileLabel, changeSelectedContent, downloadFile } from '../actions/content';
-
+import { selectFile, unselectAllJobFiles, unselectAllJobs, highlightSelected } from '../actions/jobNodes';
 
 class JobFile extends React.Component {
     constructor(props) {
@@ -32,6 +32,14 @@ class JobFile extends React.Component {
             menuShortCuts: true,
             menuVisible: false,
         };
+        this.checkForShowDD();
+    }
+
+    checkForShowDD() {
+        const { file, showDD } = this.props;
+        if (file && file.label && file.label === showDD) {
+            this.openFile();
+        }
     }
 
     isFileOpen() {
@@ -51,6 +59,9 @@ class JobFile extends React.Component {
         } else {
             dispatch(fetchJobFile(job.get('jobName'), job.get('jobId'), file.label, file.id));
         }
+        dispatch(unselectAllJobFiles());
+        dispatch(unselectAllJobs());
+        dispatch(selectFile(job.get('jobId'), file.label));
     }
 
     refreshFile() {
@@ -74,6 +85,13 @@ class JobFile extends React.Component {
     hideContextMenu() {
         hideMenu();
         this.setState({ menuVisible: false });
+    }
+
+    handleContextMenu() {
+        const { dispatch, file } = this.props;
+        if (file.selectionType !== 'selected') {
+            dispatch(highlightSelected());
+        }
     }
 
     handleKeyDown(e) {
@@ -143,11 +161,16 @@ class JobFile extends React.Component {
                         <span
                             className="content-link"
                             onClick={() => { this.openFile(); }}
+                            onContextMenu={() => { this.handleContextMenu(); }}
                             onKeyDown={this.handleKeyDown}
                             tabIndex="0"
                             role="treeitem"
                             aria-level="2"
                             aria-haspopup={true}
+                            style={this.state.menuVisible ? { border: '1px solid #333333' }
+                                : file.selectionType === 'selected' ? { background: '#dedede', border: '1px solid #333333' }
+                                    : file.selectionType === 'highlighted' ? { background: '#dedede' }
+                                        : null}
                         >
                             <Description className="node-icon" />
                             <span className="job-file-label">{file.label}</span>
@@ -161,8 +184,10 @@ class JobFile extends React.Component {
 
 JobFile.propTypes = {
     job: PropTypes.instanceOf(Map).isRequired,
+    showDD: PropTypes.string,
     content: PropTypes.instanceOf(List),
     file: PropTypes.shape({
+        selectionType: PropTypes.string.isRequired,
         label: PropTypes.string.isRequired,
         id: PropTypes.number.isRequired,
     }),