+++ WebContent/js/components/JobInstance.js
@@ -15,7 +15,7 @@ import { connect } from 'react-redux';
 import LabelIcon from '@material-ui/icons/Label';
 import { ContextMenu, MenuItem, ContextMenuTrigger } from 'react-contextmenu';
 import { hideMenu } from 'react-contextmenu/modules/actions';
-import { fetchJobFiles, toggleJob, invertJobSelectStatus, unselectAllJobs, cancelJob, purgeJob, purgeJobs, getSelectedJobs } from '../actions/jobNodes';
+import { fetchJobFiles, toggleJob, invertJobSelectStatus, unselectAllJobs, cancelJob, purgeJob, purgeJobs, getSelectedJobs, unselectAllJobFiles, highlightSelected } from '../actions/jobNodes';
 import { getJCL, getFileLabel, changeSelectedContent, fetchConcatenatedJobFiles, downloadAllJobFiles, downloadFile } from '../actions/content';
 import JobFile from './JobFile';
 import JobStep from './JobStep';
@@ -35,6 +35,15 @@ class JobInstance extends React.Component {
         };
 
         this.handleKeyDown = this.handleKeyDown.bind(this);
+        this.checkForExpand = this.checkForExpand.bind(this);
+        setTimeout(this.checkForExpand, 1500); // Give time for UI code to render before toggling expand
+    }
+
+    checkForExpand() {
+        const { expand, pos } = this.props;
+        if (typeof (expand) === 'boolean' && expand && pos === 0) {
+            this.handleJobToggle();
+        }
     }
 
     handleSingleClick(e) {
@@ -43,6 +52,7 @@ class JobInstance extends React.Component {
             this.handleSelectChange();
         } else {
             dispatch(unselectAllJobs());
+            dispatch(unselectAllJobFiles());
             this.handleSelectChange();
             this.state.singleClickTimeout = setTimeout(() => {
                 if (!this.state.preventSingleClick) {
@@ -59,6 +69,13 @@ class JobInstance extends React.Component {
         this.setState({ menuVisible: false });
     }
 
+    handleContextMenu() {
+        const { dispatch, job } = this.props;
+        if (job.get('selectionType') !== 'selected') {
+            dispatch(highlightSelected());
+        }
+    }
+    
     handleKeyDown(e) {
         const { job } = this.props;
         const shortCuts = ['o', 'j', 'c', 'delete', 'r', 'd'];
@@ -219,10 +236,19 @@ class JobInstance extends React.Component {
     }
 
     renderJobFiles() {
-        const { job, dispatch } = this.props;
+        const { job, dispatch, showDD } = this.props;
         const files = job.get('files');
         return files.map(file => {
-            return (<JobFile key={file.id} job={job} dispatch={dispatch} file={file} />);
+            return (<JobFile
+                key={file.id}
+                job={job}
+                showDD={showDD}
+                dispatch={dispatch}
+                file={file}
+                style={file.selectionType === 'selected' ? { background: '#dedede', border: '1px solid #333333' }
+                    : file.selectionType === 'highlighted' ? { background: '#dedede', border: null } 
+                        : null}
+            />);
         });
     }
 
@@ -285,13 +311,13 @@ class JobInstance extends React.Component {
             <div
                 className="job-instance"
                 role="none"
-                style={job.get('isSelected') ? { background: '#dedede' } : null}
             >
                 <li role="none">
                     <ContextMenuTrigger id={job.get('label')}>
                         <span
                             className="content-link"
                             onClick={e => { this.handleSingleClick(e); }}
+                            onContextMenu={e => { this.handleContextMenu(e); }}
                             onDoubleClick={() => { this.handleOpenAllFiles(); }}
                             tabIndex="0"
                             onKeyDown={this.handleKeyDown}
@@ -299,6 +325,10 @@ class JobInstance extends React.Component {
                             aria-expanded={job.get('isToggled').toString()}
                             aria-level="1"
                             aria-haspopup={true}
+                            style={this.state.menuVisible ? { border: '1px solid #333333' }
+                                : job.get('selectionType') === 'selected' ? { background: '#dedede', border: '1px solid #333333' }
+                                    : job.get('selectionType') === 'highlighted' ? { background: '#dedede', border: null }
+                                        : null}
                         >
                             <LabelIcon className="node-icon" />
                             <span className="job-label">
@@ -307,7 +337,10 @@ class JobInstance extends React.Component {
                             </span>
                         </span>
                     </ContextMenuTrigger>
-                    <ul role="group">
+                    <ul
+                        role="group"
+                        style={{ background: '#F5F8F8' }}
+                    >
                         {job.get('isToggled') && this.renderJobFiles(job)}
                     </ul>
                 </li>
@@ -318,6 +351,9 @@ class JobInstance extends React.Component {
 
 JobInstance.propTypes = {
     dispatch: PropTypes.func.isRequired,
+    showDD: PropTypes.string,
+    expand: PropTypes.bool,
+    pos: PropTypes.number,
     job: PropTypes.instanceOf(Map).isRequired,
     jobs: PropTypes.instanceOf(List).isRequired,
     content: PropTypes.instanceOf(List),
@@ -325,10 +361,12 @@ JobInstance.propTypes = {
 
 function mapStateToProps(state) {
     const contentRoot = state.get('content');
+    const filtersRoot = state.get('filters');
     const jobs = state.get('jobNodes');
     return {
         content: contentRoot.get('content'),
         jobs: jobs.get('jobs'),
+        showDD: filtersRoot.get('showDD'),
     };
 }