+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the JES-Explorer will be documented in this file.
 
 ### Bug fixes
 - Fixed bug where URL requests fail when using '#' character in prefix.
+- Fixed a bug where using app2app params at launch would not use desired data
+
+### New features and enhancements
+- Highlight the Selected Jobs and Job Files
+- Update the Job tree when a job is deleted or cancelled
+- Added '*' support for job id
+- Added app2app arguments: 'expand' - boolean that says to expand the job. In a list of jobs, this expands the first result. 'showDD' - string that will auto-open any dataset definition with this name, when expanding job.
 
 ## <2.0.0>